Create UnlinkClassCommand
There is currently no way to unlink a student or tutor from a class in
the app.

Being able to unlink (i.e. delete the link) is essential for the app to
be fully functional.

Create an 'unlinkClass' command that functions similarly to 'linkClass',
except it removes the link instead of creating it, between a specified
class and a student/tutor.