Bug fix: non-alphanumeric inputs for commands
When getClasses/getParent/getStudents is run with non-alphanumeric
characters in the n/NAME parameter, the app will not respond and the
command box does not update (when it usually displays an error message
of some kind to let the user know why the command didn't run).

With no feedback to the user on why the command fails to run, it makes
it hard for users to fix the issue and continue to use the app smoothly.

Let's update the relevant parser classes to properly catch the
IllegalArgumentException and convert it to a ParseException, so that the
error logic of our code can handle it as expected.

Throughout our code, the ParseException is used to contain any errors
with parsing, attaching the relevant error message to it, for it to be
caught by the main logic and processed to be displayed in the GUI. The
main logic, however, is not designed to catch IllegalArgumentExceptions,
so it gets caught by the backend instead and does not update the GUI in
any way.

The IllegalArgumentException is thrown by the attempt to create a new
Name object in the parser classes. More specifically, the constructor
method for Name calls AppUtils::checkArgument, which takes in a boolean
(where the coder can input any conditional that evaluates to a boolean)
and throws the IllegalArgumentException if false.