Bug fix: incomplete linkage updates with edit
There are a bunch of issues that stem from the fact that the edit
command does not properly update all linkages when the category of a
Person is changed. For example, when a tutor is changed to a student,
the classes that the tutor is assigned to should have the tutor's ID
removed from the class, but it does not happen in the current codebase.

This bug can cause severe issues with data correctness and consistency
throughout the app. The known bugs that stem from this issue include:
- not being able to assign a new tutor to a class after changing the
category of the old tutor.
- a student remaining in the class list of a class after having its
category changed, which if they are changed back to a student in the
future, will show up in the class as a student again, even though that
link should have been removed when the student's category is changed.
- a student remaining in the class list of a class after having its
category changed, which if they are changed back to a student in the
future and assigned to a different class, allows them to be assigned to
more than one class.

Let's update EditCommand so that when the category of a person is
changed, all linkages are updated accordingly.