Bug fix: linking issue with same name persons
If there is a student and a tutor of the exact same name, any attempts
to link/unlink them to/from a class by name will only ever be able to
link/unlink the person who is higher up on the list.

This makes the second person fully inaccessible when it comes to linking
and unlinking classes, which is a functional issue.

Let's add an optional parameter to specify the category of the person
we want to link/unlink.

This is sufficient as the constraints of TutorFlow assert that there can
be at most three Persons of the same name, where all three are of
different categories (i.e. student, parent, tutor). Specifying the
category on top of the name will ensure the command can find a unique
person (if there is any).